requirements.txt
fastapi==0.115.11
psycopg2-binary
sqlalchemy==2.0.39
pydantic==2.10.6
python-jose==3.3.0
python-dotenv==1.0.1
faker==30.3.0
bcrypt==4.0.1
uvicorn==0.34.0
boto3
aiofiles
python-docx
PyPDF2
python-pptx
playwright
openai